backend/requirements.txt
fastapi==0.104.1
uvicorn==0.22.0
pydantic==2.4.2
PyJWT==2.8.0
pyyaml
pymongo
scipy